+++ src/SMAPI.Web/Framework/Caching/BaseCacheRepository.cs
@@ -10,10 +10,10 @@ namespace StardewModdingAPI.Web.Framework.Caching
         *********/
         /// <summary>Whether cached data is stale.</summary>
         /// <param name="lastUpdated">The date when the data was updated.</param>
-        /// <param name="cacheMinutes">The age in minutes before data is considered stale.</param>
-        public bool IsStale(DateTimeOffset lastUpdated, int cacheMinutes)
+        /// <param name="staleMinutes">The age in minutes before data is considered stale.</param>
+        public bool IsStale(DateTimeOffset lastUpdated, int staleMinutes)
         {
-            return lastUpdated < DateTimeOffset.UtcNow.AddMinutes(-cacheMinutes);
+            return lastUpdated < DateTimeOffset.UtcNow.AddMinutes(-staleMinutes);
         }
     }
 }